requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
pillow>=8.0.0
numpy>=1.20.0
huggingface_hub>=0.10.0
safetensors>=0.3.0